+++ lib/govspeak.rb
@@ -81,14 +81,6 @@ module Govspeak
       ERB::Util.html_escape(string || "").strip.gsub(/(?:\r?\n)/, "<br/>").html_safe
     end
 
-    def to_sanitized_html
-      HtmlSanitizer.new(to_html).sanitize
-    end
-
-    def to_sanitized_html_without_images
-      HtmlSanitizer.new(to_html).sanitize_without_images
-    end
-
     def to_text
       HTMLEntities.new.decode(to_html.gsub(/(?:<[^>]+>|\s)+/, " ").strip)
     end