+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 * Allow sanitize option on remove invalid HTML from source input
+* BREAKING CHANGE: Remove `to_sanitized_html` method in favour of `sanitize` option on initialize
+* BREAKING CHANGE: Remove `to_sanitized_html_without_images` as no apps use this anymore
 
 ## 5.9.1